# nodejs-server/api/openapi.core.yaml
openapi: 3.0.3
info:
  title: Argovis API
  description: "REST API for Argo profiles, platforms, selections, BGC data, gridded\
    \ products, etc."
  contact:
    name: Argovis Collaboration
    url: https://argovis.colorado.edu
    email: argovis@colorado.edu
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: "2.0"
...
servers:
- url: https://argovis-api.colorado.edu
- url: /
security:
- name: []
...
paths:
...
  /grids/{gridName}:
    get:
      tags:
      - grid
      summary: Search and filter for grid named in path
      operationId: findgrid
      parameters:
      - name: gridName
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          enum:
          - rg09
          - kg21
          - glodap
      - name: id
        in: query
        description: Unique ID to search for.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: 4902911_0
      - name: startDate
        in: query
        description: ISO 8601 UTC date-time formatted string indicating the beginning
          of the time period of interest.
        required: false
        style: form
        explode: true
        allowReserved: true
        schema:
          type: string
          format: date-time
          example: 2010-01-01T00:00:00Z
      - name: endDate
        in: query
        description: ISO 8601 UTC date-time formatted string indicating the end of
          the time period of interest.
        required: false
        style: form
        explode: true
        allowReserved: true
        schema:
          type: string
          format: date-time
          example: 2010-01-01T00:00:00Z
      - name: polygon
        in: query
        description: "array of [lon, lat] vertices describing a polygon bounding the\
          \ region of interest; final point must match initial point"
        required: false
        style: form
        explode: true
        allowReserved: true
        schema:
          type: string
          example: "[[-74.1,40.5],[-73.7,40.5],[-74.0,40.8],[-74.1,40.5]]"
      - name: multipolygon
        in: query
        description: array of polygon regions; region of interest is taken as the
          intersection of all listed polygons.
        required: false
        style: form
        explode: true
        allowReserved: true
        schema:
          type: string
          example: "[[[0,0],[0,1],[1,1],[1,0],[0,0]],[[0.5,0],[0.5,1],[1.5,1],[1.5,0],[0.5,0]]]"
      - name: winding
        in: query
        description: Enforce ccw winding for polygon and multipolygon
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - "true"
      - name: center
        in: query
        description: center to measure max radius from when defining circular region
          of interest; must be used in conjunction with query string parameter 'radius'.
        required: false
        style: form
        explode: false
        allowReserved: true
        schema:
          type: array
          example: "10,20.1"
          items:
            maxItems: 2
            minItems: 2
            type: number
      - name: radius
        in: query
        description: km from centerpoint when defining circular region of interest;
          must be used in conjunction with query string parameter 'center'.
        required: false
        style: form
        explode: true
        schema:
          type: number
          example: 50
      - name: compression
        in: query
        description: Data minification strategy to apply.
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - minimal
      - name: mostrecent
        in: query
        description: get back only the n records with the most recent values of timestamp.
        required: false
        style: form
        explode: true
        schema:
          type: number
      - name: data
        in: query
        description: "Keys of data to include. Return only documents that have all\
          \ data requested, within the pressure range if specified. Accepts ~ negation\
          \ to filter out documents including the specified data. Omission of this\
          \ parameter will result in metadata only responses."
        required: false
        style: form
        explode: false
        allowReserved: true
        schema:
          type: array
          example: rg09_salinity
          items:
            type: string
      - name: presRange
        in: query
        description: Pressure range in dbar to filter for; levels outside this range
          will not be returned.
        required: false
        style: form
        explode: false
        allowReserved: true
        schema:
          type: array
          example: "10,20.1"
          items:
            maxItems: 2
            minItems: 2
            type: number
      - name: batchmeta
        in: query
        description: return the metadata documents corresponding to a temporospatial
          data search
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/grid'
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      x-swagger-router-controller: Grid
  /grids/meta:
    get:
      tags:
      - grid
      summary: Metadata for grids by ID
      operationId: findgridMeta
      parameters:
      - name: id
        in: query
        description: Unique ID to search for.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: 4902911_0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/gridMeta'
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      x-swagger-router-controller: Grid
  /grids/{gridName}/vocabulary:
    get:
      tags:
      - grid
      summary: List data and lattice for the requested grid.
      operationId: gridVocab
      parameters:
      - name: gridName
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          enum:
          - rg09
          - kg21
          - glodap
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      x-swagger-router-controller: Grid
...
  /extended/{extendedName}:
    get:
      tags:
      - extended
      summary: Search and filter for extended object named in path
      operationId: findExtended
      parameters:
      - name: extendedName
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          enum:
          - ar
      - name: id
        in: query
        description: Unique ID to search for.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: 4902911_0
      - name: startDate
        in: query
        description: ISO 8601 UTC date-time formatted string indicating the beginning
          of the time period of interest.
        required: false
        style: form
        explode: true
        allowReserved: true
        schema:
          type: string
          format: date-time
          example: 2010-01-01T00:00:00Z
      - name: endDate
        in: query
        description: ISO 8601 UTC date-time formatted string indicating the end of
          the time period of interest.
        required: false
        style: form
        explode: true
        allowReserved: true
        schema:
          type: string
          format: date-time
          example: 2010-01-01T00:00:00Z
      - name: polygon
        in: query
        description: "array of [lon, lat] vertices describing a polygon bounding the\
          \ region of interest; final point must match initial point"
        required: false
        style: form
        explode: true
        allowReserved: true
        schema:
          type: string
          example: "[[-74.1,40.5],[-73.7,40.5],[-74.0,40.8],[-74.1,40.5]]"
      - name: multipolygon
        in: query
        description: array of polygon regions; region of interest is taken as the
          intersection of all listed polygons.
        required: false
        style: form
        explode: true
        allowReserved: true
        schema:
          type: string
          example: "[[[0,0],[0,1],[1,1],[1,0],[0,0]],[[0.5,0],[0.5,1],[1.5,1],[1.5,0],[0.5,0]]]"
      - name: winding
        in: query
        description: Enforce ccw winding for polygon and multipolygon
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - "true"
      - name: center
        in: query
        description: center to measure max radius from when defining circular region
          of interest; must be used in conjunction with query string parameter 'radius'.
        required: false
        style: form
        explode: false
        allowReserved: true
        schema:
          type: array
          example: "10,20.1"
          items:
            maxItems: 2
            minItems: 2
            type: number
      - name: radius
        in: query
        description: km from centerpoint when defining circular region of interest;
          must be used in conjunction with query string parameter 'center'.
        required: false
        style: form
        explode: true
        schema:
          type: number
          example: 50
      - name: compression
        in: query
        description: Data minification strategy to apply.
        required: false
        style: form
        explode: true
        schema:
          type: string
          enum:
          - minimal
      - name: mostrecent
        in: query
        description: get back only the n records with the most recent values of timestamp.
        required: false
        style: form
        explode: true
        schema:
          type: number
      - name: batchmeta
        in: query
        description: return the metadata documents corresponding to a temporospatial
          data search
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/genericExtended'
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      x-swagger-router-controller: Extended
  /extended/meta:
    get:
      tags:
      - extended
      summary: Metadata for extended objects by ID
      operationId: findextendedMeta
      parameters:
      - name: id
        in: query
        description: Unique ID to search for.
        required: false
        style: form
        explode: true
        schema:
          type: string
          example: 4902911_0
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/genericExtendedMeta'
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      x-swagger-router-controller: Extended
  /extended/{extendedName}/vocabulary:
    get:
      tags:
      - extended
      summary: Vocab data for the named extended object.
      operationId: extendedVocab
      parameters:
      - name: extendedName
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
          enum:
          - ar
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                x-content-type: application/json
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
        "500":
          description: Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorResponse'
      x-swagger-router-controller: Extended
...
components:
  schemas:
...
    grid:
      required:
      - _id
      - basin
      - geolocation
      - metadata
      - timestamp
      type: object
      properties:
        _id:
          type: string
        metadata:
          type: array
          items:
            type: string
        geolocation:
          $ref: '#/components/schemas/geoJSONpoint'
        basin:
          type: number
        timestamp:
          type: string
          format: date-time
        data:
          $ref: '#/components/schemas/data'
      example:
        metadata:
        - metadata
        - metadata
        data:
        - - ""
          - ""
        - - ""
          - ""
        _id: _id
        basin: 0.80082819046101150206595775671303272247314453125
        geolocation:
          coordinates:
          - 0.80082819046101150206595775671303272247314453125
          - 0.80082819046101150206595775671303272247314453125
          type: type
        timestamp: 2000-01-23T04:56:07.000+00:00
    gridMeta:
      required:
      - _id
      - data_info
      - data_type
      - date_updated_argovis
      - levels
      - source
      type: object
      properties:
        _id:
          type: string
        data_type:
          type: string
        date_updated_argovis:
          type: string
          format: date-time
        source:
          type: array
          items:
            $ref: '#/components/schemas/generic_source_info'
        levels:
          type: array
          items:
            type: number
        data_info:
          $ref: '#/components/schemas/data_info'
      example:
        data_info:
        - ""
        - ""
        data_type: data_type
        _id: _id
        source:
        - date_updated: 2000-01-23T04:56:07.000+00:00
          source:
          - source
          - source
          url: url
          doi: doi
        - date_updated: 2000-01-23T04:56:07.000+00:00
          source:
          - source
          - source
          url: url
          doi: doi
        date_updated_argovis: 2000-01-23T04:56:07.000+00:00
        levels:
        - 0.80082819046101150206595775671303272247314453125
        - 0.80082819046101150206595775671303272247314453125
    geoJSONpoint:
      required:
      - coordinates
      - type
      type: object
      properties:
        type:
          type: string
        coordinates:
          type: array
          items:
            maxItems: 2
            minItems: 2
            type: number
      example:
        coordinates:
        - 0.80082819046101150206595775671303272247314453125
        - 0.80082819046101150206595775671303272247314453125
        type: type
...
    geoJSONmultipolygon:
      type: object
      properties:
        type:
          type: string
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                type: array
                items:
                  maxItems: 2
                  minItems: 2
                  type: number
      example:
        coordinates:
        - - - - 0.80082819046101150206595775671303272247314453125
              - 0.80082819046101150206595775671303272247314453125
            - - 0.80082819046101150206595775671303272247314453125
              - 0.80082819046101150206595775671303272247314453125
          - - - 0.80082819046101150206595775671303272247314453125
              - 0.80082819046101150206595775671303272247314453125
            - - 0.80082819046101150206595775671303272247314453125
              - 0.80082819046101150206595775671303272247314453125
        - - - - 0.80082819046101150206595775671303272247314453125
              - 0.80082819046101150206595775671303272247314453125
            - - 0.80082819046101150206595775671303272247314453125
              - 0.80082819046101150206595775671303272247314453125
          - - - 0.80082819046101150206595775671303272247314453125
              - 0.80082819046101150206595775671303272247314453125
            - - 0.80082819046101150206595775671303272247314453125
              - 0.80082819046101150206595775671303272247314453125
        type: type
    data:
      type: array
      items:
        type: array
        items:
          anyOf:
          - type: string
          - type: number
    data_info:
      type: array
      items:
        anyOf:
        - type: array
          items:
            type: string
        - type: array
          items:
            type: array
            items:
              anyOf:
              - type: string
              - type: number
    generic_source_info:
      type: object
      properties:
        source:
          type: array
          items:
            type: string
        url:
          type: string
        date_updated:
          type: string
          format: date-time
        doi:
          type: string
      example:
        date_updated: 2000-01-23T04:56:07.000+00:00
        source:
        - source
        - source
        url: url
        doi: doi
...
    genericExtended:
      required:
      - _id
      - basins
      - geolocation
      - metadata
      - raster
      - timestamp
      type: object
      properties:
        _id:
          type: string
        metadata:
          type: array
          items:
            type: string
        geolocation:
          $ref: '#/components/schemas/geoJSONmultipolygon'
        true_geolocation:
          $ref: '#/components/schemas/geoJSONmultipolygon'
        timestamp:
          type: string
          format: date-time
        raster:
          type: array
          items:
            type: array
            items:
              anyOf:
              - type: number
              - type: array
                items:
                  type: number
        basins:
          type: array
          items:
            type: integer
        flags:
          type: array
          items:
            type: string
      example:
        metadata:
        - metadata
        - metadata
        true_geolocation: null
        raster:
        - - ""
          - ""
        - - ""
          - ""
        flags:
        - flags
        - flags
        _id: _id
        basins:
        - 6
        - 6
        geolocation:
          coordinates:
          - - - - 0.80082819046101150206595775671303272247314453125
                - 0.80082819046101150206595775671303272247314453125
              - - 0.80082819046101150206595775671303272247314453125
                - 0.80082819046101150206595775671303272247314453125
            - - - 0.80082819046101150206595775671303272247314453125
                - 0.80082819046101150206595775671303272247314453125
              - - 0.80082819046101150206595775671303272247314453125
                - 0.80082819046101150206595775671303272247314453125
          - - - - 0.80082819046101150206595775671303272247314453125
                - 0.80082819046101150206595775671303272247314453125
              - - 0.80082819046101150206595775671303272247314453125
                - 0.80082819046101150206595775671303272247314453125
            - - - 0.80082819046101150206595775671303272247314453125
                - 0.80082819046101150206595775671303272247314453125
              - - 0.80082819046101150206595775671303272247314453125
                - 0.80082819046101150206595775671303272247314453125
          type: type
        timestamp: 2000-01-23T04:56:07.000+00:00
    genericExtendedMeta:
      required:
      - _id
      - data_info
      - data_type
      - date_updated_argovis
      - source
      type: object
      properties:
        _id:
          type: string
        data_type:
          type: string
        data_info:
          $ref: '#/components/schemas/data_info'
        date_updated_argovis:
          type: string
          format: date-time
        source:
          type: array
          items:
            $ref: '#/components/schemas/generic_source_info'
      example:
        data_info:
        - ""
        - ""
        data_type: data_type
        _id: _id
        source:
        - date_updated: 2000-01-23T04:56:07.000+00:00
          source:
          - source
          - source
          url: url
          doi: doi
        - date_updated: 2000-01-23T04:56:07.000+00:00
          source:
          - source
          - source
          url: url
          doi: doi
        date_updated_argovis: 2000-01-23T04:56:07.000+00:00
    errorResponse:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
...